requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
aiomysql  # Replace with asyncpg for PostgreSQL
#asyncpg # For PGSQL
pydantic
python-jose
passlib
alembic
psycopg2